+++ server/app/Http/Controllers/PurchaseReturnPaymentController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Helpers\Constants;
+use Illuminate\Http\Request;
+use App\Models\PurchaseReturn;
 use App\Models\PurchaseReturnPayment;
 use App\Requests\PurchaseReturnPaymentRequest;
-use Illuminate\Http\Request;
 
 class PurchaseReturnPaymentController extends Controller
 {
@@ -12,15 +14,19 @@ class PurchaseReturnPaymentController extends Controller
     {
         $attr = PurchaseReturnPaymentRequest::validationCreate($req);
 
-        $purchase = PurchaseReturnPayment::find($purchaseId);
+        $purchase = PurchaseReturn::find($purchaseId);
 
         if (!$purchase) return $this->error('The purchase return invoice is not found to add payment', 404);
 
-        if ($purchase->payment_status == 'Paid') {
-            return $this->error('This purchase return invoice has been paid !', 422);
+        if ($purchase->payment_status == Constants::PAYMENT_STATUS_PAID) {
+          return $this->error('This purchase invoice has been paid !', 422);
         }
 
         PurchaseReturnPayment::create($attr);
+
+        $purchase->payment_status = $purchase->new_payment_status;
+    
+        $purchase->save();
 
         return $this->success([], 'The Payment has been added successfully');
     }
@@ -29,6 +35,10 @@ class PurchaseReturnPaymentController extends Controller
     public function update(Request $req, $purchaseId, $id)
     {
         $attr = PurchaseReturnPaymentRequest::validationUpdate($req);
+
+        $purchase = PurchaseReturn::find($purchaseId);
+    
+        if (!$purchase) return $this->error('The purchase invoice is not found to add payment', 404);
 
         $payment = PurchaseReturnPayment::find($id);
 
@@ -40,6 +50,10 @@ class PurchaseReturnPaymentController extends Controller
 
         $payment->save();
 
+        $purchase->payment_status = $purchase->new_payment_status;
+    
+        $purchase->save();
+
         return $this->success([], 'The Payment has been updated successfully');
     }
 
@@ -48,11 +62,19 @@ class PurchaseReturnPaymentController extends Controller
     {
         PurchaseReturnPaymentRequest::validationRemove($req);
 
+        $purchase = PurchaseReturn::find($purchaseId);
+    
+        if (!$purchase) return $this->error('The purchase invoice is not found to add payment', 404);
+
         $payment = PurchaseReturnPayment::find($id);
 
         if (!$payment) return $this->error('This payment is not found', 404);
 
         $payment->delete();
+
+        $purchase->payment_status = $purchase->new_payment_status;
+    
+        $purchase->save();
 
         return $this->success([], 'The Payment has been deleted successfully');
     }
